+++ src/RGBdistance.ts
@@ -1,5 +1,5 @@
 import { ciede2000 } from "./CIEDE2000.js";
-import { convertRGBtoLab } from "./converters.js";
+import { convertRGBtoLab } from "./convertRGBtoLab.js";
 import type { RGBColour } from "./types";
 
 export const RGBdistance = (colour1: RGBColour, colour2: RGBColour): number => {
