+++ src/index.ts
@@ -1,19 +1,16 @@
-export { ciede2000 } from "./CIEDE2000.js";
-export {
-  convertCSSRGBtoHex,
-  convertHextoRGB,
-  convertNamedCSSColourtoHex,
-  convertRGBtoHex,
-  convertRGBtoLab,
-  convertRGBtoXYZ,
-  convertXYZtoLab,
-} from "./converters.js";
 export { RGBdistance } from "./RGBdistance.js";
+export { ciede2000 } from "./CIEDE2000.js";
+export { convertCSSRGBtoHex } from "./convertCSSRGBtoHex.js";
+export { convertHextoRGB } from "./convertHextoRGB.js";
+export { convertNamedCSSColourtoHex } from "./convertNamedCSSColourtoHex.js";
+export { convertRGBtoHex } from "./convertRGBtoHex.js";
+export { convertRGBtoLab } from "./convertRGBtoLab.js";
+export { convertXYZtoLab } from "./convertXYZtoLab.js";
 export { getSRGBLuminanceFromHex } from "./getSRGBLuminanceFromHex.js";
 export { getSRGBLuminanceFromRGB } from "./getSRGBLuminanceFromRGB.js";
-export { pickHexColourContrast } from "./pickHexColourContrast.js";
-export { isCSSRGBColour } from "./assertions.js";
 export { isCSSNamedDarkColour } from "./isCSSNameDarkColour.js";
+export { isCSSRGBColour } from "./assertions.js";
 export { isCSSRGBDarkColour } from "./isCSSRGBDarkColour.js";
 export { isDarkColour } from "./isDarkColour.js";
 export { isHexDarkColour } from "./isHexDarkColour.js";
+export { pickHexColourContrast } from "./pickHexColourContrast.js";
